+++ src/main/java/com/tiansi/annotation/service/ClipsService.java
@@ -6,12 +6,12 @@ import com.tiansi.annotation.domain.Clips;
 import java.util.List;
 
 public interface ClipsService extends IService<Clips> {
-    List<Clips> findByVideoId(int videoId);
-    List<Clips> findByVideoIdUntagged(int videoId);
+    List<Clips> findByVideoId(Long videoId);
+    List<Clips> findByVideoIdUntagged(Long videoId);
     List<Clips> findAll();
     List<Clips> findAllUntagged();
-    boolean deleteByVideoId(int videoId);
+    boolean deleteByVideoId(Long videoId);
     boolean tag(Clips clips);
-    boolean clear(int id);
-    int clearBatch(List<Integer> ids);
+    boolean clear(Long id);
+    int clearBatch(List<Long> ids);
 }
