+++ src/main/java/com/tiansi/annotation/service/TrialService.java
@@ -4,5 +4,5 @@ import com.baomidou.mybatisplus.extension.service.IService;
 import com.tiansi.annotation.domain.Trial;
 
 public interface TrialService extends IService<Trial> {
-    int insert(Trial trial);
+    Long insert(Trial trial);
 }
